rename to rippled-json-rpc-schema
